Fix Storybook build errors by separating display components into single files.
