+++ package.json
@@ -31,6 +31,7 @@
     "bulma": "^0.6.2",
     "classnames": "^2.2.6",
     "concurrently": "^5.0.0",
+    "fs": "^0.0.1-security",
     "gatsby": "^2.17.2",
     "gatsby-plugin-google-tagmanager": "^2.1.15",
     "gatsby-plugin-load-script": "^1.0.6",
