+++ src/cms/preview-templates/FooterPreview.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent, ComponentProps } from 'react'
 import '../../scss/main.scss'
 import { PreviewProps, getJSON } from './helpers'
 import * as R from 'ramda'
-import { FooterData, FooterDataOnDisk } from '../../data/footer'
+import { FooterData, FooterDataOnDisk } from '../../data/footer/types'
 import { FooterComponent } from '../../components/Footer/component'
 import { SafeExternalLink } from '../../components/Link'
 import { createPreview } from './Preview'
