+++ src/cms/preview-templates/HomePreview.tsx
@@ -1,10 +1,10 @@
 import React, { FunctionComponent } from 'react'
 import { PreviewProps, getJSON } from './helpers'
-import { HomePageComponent } from '../../components/Home'
-import { HomeData, HomeDataOnDisk } from '../../data/home'
+import { HomePageComponent } from '../../components/Home/component'
+import { HomeData, HomeDataOnDisk } from '../../data/home/types'
 import { createPreview } from './Preview'
 import { filterByLocale } from '../../data/i18n'
-import { ProjectData } from '../../data/project'
+import { ProjectData } from '../../data/project/types'
 
 const Preview = createPreview<
   HomeDataOnDisk,
