+++ src/cms/preview-templates/ImprintPreview.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react'
-import { ImprintComponent } from '../../components/Imprint'
-import { Imprint, ImprintPageOnDisk } from '../../data/imprint'
+import { ImprintComponent } from '../../components/Imprint/component'
+import { Imprint, ImprintPageOnDisk } from '../../data/imprint/types'
 import { PreviewProps, getJSON } from './helpers'
 import { filterByLocale } from '../../data/i18n'
 import { createPreview } from './Preview'
