+++ src/cms/preview-templates/NavbarPreview.tsx
@@ -1,12 +1,16 @@
 import React, { FunctionComponent } from 'react'
 import '../../scss/main.scss'
 import { PreviewProps, getJSON } from './helpers'
-import { TranslationCollection, Locale, filterByLocale } from '../../data/i18n'
+import { filterByLocale } from '../../data/i18n'
 import * as R from 'ramda'
-import { Navbar as NavbarData } from '../../data/navbar'
+import { Navbar as NavbarData } from '../../data/navbar/types'
 import { SafeExternalLink } from '../../components/Link'
 import { createPreview } from './Preview'
-import { NavbarComponent, NavbarComponentProps } from '../../components/Navbar'
+import { TranslationCollection, Locale } from '../../data/types'
+import {
+  NavbarComponent,
+  NavbarComponentProps,
+} from '../../components/Navbar/component'
 
 const Preview = createPreview<
   TranslationCollection<NavbarData>,
