+++ src/components/Footer/index.tsx
@@ -3,9 +3,9 @@ import { FooterComponent } from './component'
 import { FunctionComponent } from 'react'
 import { Link } from 'gatsby'
 import { SafeExternalLink } from '../Link'
-import { FooterData } from '../../data/footer'
 import { usePathPrefix } from '../../utils/hooks'
 import { useFooter } from '../../hooks/data'
+import { FooterData } from '../../data/footer/types'
 
 export const Footer: FunctionComponent = () => {
   const footerData: FooterData = useFooter()
