+++ src/components/Imprint/index.tsx
@@ -1,22 +1,7 @@
 import React, { FunctionComponent } from 'react'
-import { TitleBanner } from '../TitleBanner/index'
-import { ContentWrap } from '../Content/index'
-import { Imprint as ImprintData } from '../../data/imprint'
-import ReactMarkdown from 'react-commonmark'
+import { Imprint as ImprintData } from '../../data/imprint/types'
 import { useImprint } from '../../hooks/data'
-
-export const ImprintComponent: FunctionComponent<ImprintData> = ({
-  content,
-  headline,
-  subheadline,
-}) => (
-  <article>
-    <TitleBanner title={headline} subtitle={subheadline} />
-    <ContentWrap>
-      <ReactMarkdown source={content} />
-    </ContentWrap>
-  </article>
-)
+import { ImprintComponent } from './component'
 
 export const Imprint: FunctionComponent<{}> = () => {
   const imprint: ImprintData = useImprint()
