+++ src/components/TitleBanner/index.tsx
@@ -1,10 +1,7 @@
 import React, { FunctionComponent, useEffect } from 'react'
-import { Container } from 'reactstrap'
 import useIsInViewport from 'use-is-in-viewport'
 import { useNavbarState } from '../../utils/hooks'
-import cn from 'classnames'
-import { blurryBackgroundBanner } from '../../scss/global.module.scss'
-import * as styles from './component.module.scss'
+import { TitleBannerComponent } from './component'
 
 export const TitleBanner: FunctionComponent<{
   title: string
@@ -23,15 +20,3 @@ export const TitleBanner: FunctionComponent<{
     </div>
   )
 }
-
-export const TitleBannerComponent: FunctionComponent<{
-  title: string
-  subtitle: string
-}> = ({ title, subtitle }) => (
-  <div className={cn(styles.viewportHeight50, blurryBackgroundBanner)}>
-    <Container className="h-100 text-center d-flex flex-column justify-content-center">
-      <h3 className="mb-4">{title}</h3>
-      <h6>{subtitle}</h6>
-    </Container>
-  </div>
-)
