+++ src/data/faqEntry/helper.ts
@@ -1,5 +1,5 @@
 import * as R from 'ramda'
-import { FAQPage, FAQEntry } from '.'
+import { FAQPage, FAQEntry } from './types'
 
 export const hasFAQEntries: (faqPage: FAQPage) => boolean = R.pipe<
   FAQPage,
