+++ src/data/home/index.tsx
@@ -1,42 +1,7 @@
 import data from './home.json'
-import { TranslationCollection, Locale, filterByLocale } from '../i18n'
-import { Head } from '../../components/Head'
-
-interface Offering {
-  headline: string
-  image: string
-  text: string
-}
-
-export interface HomeData {
-  heroHeadline: string
-  heroSubheadline: string
-  offerings: Offering[]
-  contactHeadline: string
-  contactDescription: string
-  contactButtonText: string
-  moreLinkText: string
-  emailButton: string
-  contactEmail: string
-  projectsHeadline?: string
-  head: Head
-}
-
-export type HomeDataOnDisk = {
-  id: string
-  offerings: TranslationCollection<Offering>[]
-  contactEmail: string
-  head: TranslationCollection<Head>
-} & TranslationCollection<{
-  heroHeadline: string
-  heroSubheadline: string
-  contactHeadline: string
-  contactDescription: string
-  contactButtonText: string
-  moreLinkText: string
-  emailButton: string
-  projectsHeadline?: string
-}>
+import { filterByLocale } from '../i18n'
+import { HomeDataOnDisk, HomeData } from './types'
+import { Locale } from '../types'
 
 const home: HomeDataOnDisk = data
