+++ src/data/home/types.ts
@@ -1,0 +1,38 @@
+import { TranslationCollection } from '../types'
+import { Head } from '../../components/Head'
+
+export interface Offering {
+  headline: string
+  image: string
+  text: string
+}
+
+export interface HomeData {
+  heroHeadline: string
+  heroSubheadline: string
+  offerings: Offering[]
+  contactHeadline: string
+  contactDescription: string
+  contactButtonText: string
+  moreLinkText: string
+  emailButton: string
+  contactEmail: string
+  projectsHeadline?: string
+  head: Head
+}
+
+export type HomeDataOnDisk = {
+  id: string
+  offerings: TranslationCollection<Offering>[]
+  contactEmail: string
+  head: TranslationCollection<Head>
+} & TranslationCollection<{
+  heroHeadline: string
+  heroSubheadline: string
+  contactHeadline: string
+  contactDescription: string
+  contactButtonText: string
+  moreLinkText: string
+  emailButton: string
+  projectsHeadline?: string
+}>
