+++ src/data/imprint/index.tsx
@@ -1,22 +1,7 @@
-import { Locale, TranslationCollection, filterByLocale } from '../i18n'
+import { filterByLocale } from '../i18n'
 import data from './imprint.json'
-import { Head } from '../../components/Head'
-
-export interface Imprint {
-  headline: string
-  subheadline: string
-  content: string
-  head: Head
-}
-
-export type ImprintPageOnDisk = {
-  id: string
-  head: TranslationCollection<Head>
-} & TranslationCollection<{
-  headline: string
-  subheadline: string
-  content: string
-}>
+import { ImprintPageOnDisk, Imprint } from './types'
+import { Locale } from '../types'
 
 const imprint: ImprintPageOnDisk = data
