+++ src/data/navbar/types.ts
@@ -1,0 +1,4 @@
+export interface Navbar {
+  contactLinkText: string
+  FAQLinkText: string
+}
