+++ src/data/project/fnd.json
@@ -1,0 +1,59 @@
+{
+    "title": "FND",
+    "slug": "fnd",
+    "usedTechnologies": [
+        "{\"name\":\"ReactJS\",\"link\":\"https://reactjs.org/\"}",
+        "{\"name\":\"Typescript\",\"link\":\"https://www.typescriptlang.org/\"}"
+    ],
+    "translations": [
+        {
+            "locale": "de",
+            "value": {
+                "client": "FND Group - 01/2018 bis 10/2019",
+                "clientLink": "https://fnd.group",
+                "reference": "Florian Glatz, CEO",
+                "product": "A digital investment platform for investments in real estate funds",
+                "application": "* Landingpage with CMS and i18n (<https://fnd.group>)\n* SPA with login, identity verification (IDNow) and private key management (<https://invest.fnd.group>)\n* Smart Contract for token and token distribution",
+                "responsibilities": "* Tech department as a service\n* Planning of infrastructure\n* Planning of security model\n* Choice of technologies\n* Setup and maintenance of deployment, CI, CD, kanban board, etc.\n* Management of requirements (design and product)\n* Software development",
+                "image": "/img/mock.jpg"
+            }
+        },
+        {
+            "locale": "en",
+            "value": {
+                "client": "FND Group - 01/2018 to 10/2019",
+                "clientLink": "https://fnd.group",
+                "reference": "Florian Glatz, CEO",
+                "product": "A digital investment platform for investments in real estate funds",
+                "application": "* Landingpage with CMS and i18n (<https://fnd.group>)\n* SPA with login, identity verification (IDNow) and private key management (<https://invest.fnd.group>)\n* Smart Contract for token and token distribution",
+                "responsibilities": "* Tech department as a service\n* Planning of infrastructure\n* Planning of security model\n* Choice of technologies\n* Setup and maintenance of deployment, CI, CD, kanban board, etc.\n* Management of requirements (design and product)\n* Software development",
+                "image": "/img/mock.jpg"
+            }
+        }
+    ],
+    "head": {
+        "id": "Head",
+        "translations": [
+            {
+                "locale": "de",
+                "value": {
+                    "title": "Hardfork GmbH - Projekte",
+                    "meta": {
+                        "description": "Wir entwickeln Software nach Ihren Wünschen in Berlin",
+                        "keywords": "Softwareentwicklung"
+                    }
+                }
+            },
+            {
+                "locale": "en",
+                "value": {
+                    "title": "Hardfork GmbH - Projects",
+                    "meta": {
+                        "description": "We develop software according to your needs in Berlin, Germany",
+                        "keywords": "Software development"
+                    }
+                }
+            }
+        ]
+    }
+}
