+++ src/data/technology/types.ts
